% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accepted_name.R
\name{accepted_name}
\alias{accepted_name}
\alias{accepted_name,PostgreSQLConnection,numeric-method}
\title{Switch accepted names by synonyms in a taxon}
\usage{
\S4method{accepted_name}{PostgreSQLConnection,numeric}(taxlist, ConceptID, accepted, names2concepts, ...)
}
\arguments{
\item{taxlist}{An object of class \linkS4class{PostgreSQLConnection}.}

\item{ConceptID}{Integer containing concept IDs where to request or set names
for one category.}

\item{accepted}{Integer indicating the usage name ID to be set as accepted
name. If this usage name does not belong to the concept indicated by
'ConceptID', an error message will be retrieved.}

\item{names2concepts}{A character vector with two values, the name of the
schema and the name of the table relating names with taxon concepts.}

\item{...}{Further arguments passed among methods.}
}
\description{
If necessary to set a synonym in a taxon concept as an accepted name.
}
