% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_tax.R
\name{check_tax}
\alias{check_tax}
\alias{check_tax.PostgreSQLConnection}
\alias{check_tax,PostgreSQLConnection-method}
\title{Additional tests of consistency for taxonomic lists}
\usage{
check_tax(conn, ...)

\method{check_tax}{PostgreSQLConnection}(conn, schema = "plant_taxonomy", ...)
}
\arguments{
\item{conn}{Connection to the database as \linkS4class{PostgreSQLConnection}.}

\item{...}{Further arguments passed among methods.}

\item{schema}{Character value indicating the name of the schema containing
taxonomic information within the database.}
}
\description{
Some restrictions cannot be implemented by constraints into the database
but may affect the validity of information and exported objects.
This function carry out some additional checks to ensure consistent
taxonomic data.
}
