% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coerce-methods.R
\docType{methods}
\name{as}
\alias{as}
\alias{coerce}
\alias{coerce,lib_db,lib_df-method}
\alias{coerce,lib_df,lib_db-method}
\alias{as<-,lib_df,missing,character-method}
\alias{as<-}
\alias{as<-,lib_db,missing,character-method}
\title{Coerce objects to lists}
\usage{
\S4method{as}{lib_df,missing,character}(object, Class) <- value

\S4method{as}{lib_db,missing,character}(object, Class) <- value
}
\arguments{
\item{object}{An object to be coerced.}

\item{Class}{A character value with the target class (not used).}

\item{value}{A character value indicating the class to be coerced to. This is
only required by the replacement methods.}
}
\value{
An object of class \code{list}.
}
\description{
Coerce vegtable objects to a list with every slot as a component of the list.
This way allows to explore content and solve problems when validity checks
fail.

Coercion is applied for different classes by vegtable.
}
\author{
Miguel Alvarez \email{kamapu78@gmail.com}
}
