% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_df.R
\name{compare_df}
\alias{compare_df}
\alias{compare_df,lib_db,lib_db,missing-method}
\title{Comparison of data frames extended to electronic libraries}
\usage{
\S4method{compare_df}{lib_db,lib_db,missing}(x, y, name, ...)
}
\arguments{
\item{x}{The reference data frame.}

\item{y}{The updated data frame.}

\item{name}{Character value indicating the name of the schema in the
PostgreSQL database.}

\item{...}{Further arguments passed to \code{\link[=read_pg]{read_pg()}}.}
}
\description{
Extension of comparisons of data frames applied to relational models.
}
