% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_db.R
\name{connect_db}
\alias{connect_db}
\title{Connect to PostgreSQL database}
\usage{
connect_db(
  dbname = "",
  host = "localhost",
  port = "5432",
  user = "",
  password = "",
  pkg = "RPostgreSQL",
  ...
)
}
\arguments{
\item{dbname, host, port}{Character values passed to \code{\link[DBI:dbConnect]{DBI::dbConnect()}}.}

\item{user, password}{Character values. They are also passed to
\code{\link[DBI:dbConnect]{DBI::dbConnect()}} but can be alternatively inserted or modified in the
prompt.}

\item{pkg}{A character value indicating the alternative package used to
establish the connection. At the moment only \code{RpostgreSQL} and
\code{RPostgres} are suitable.}

\item{...}{Further arguments passed to \code{\link[DBI:dbConnect]{DBI::dbConnect()}}.}
}
\value{
A connection as \linkS4class{PostgreSQLConnection}.
}
\description{
This function opens a prompt to insert connection details to a PostgreSQL
database.
}
\references{
http://r.789695.n4.nabble.com/tkentry-that-exits-after-RETURN-tt854721.html#none

https://gist.github.com/mages/2aed2a053e355e3bfe7c#file-getlogindetails-r

Dalgaard (2001).
}
\author{
Miguel Alvarez \email{kamapu@posteo.com}
}
