% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db2vegtable.R
\name{db2vegtable}
\alias{db2vegtable}
\alias{db2vegtable.PostgreSQLConnection}
\title{Import PostgreSQL databases into vegtable objects}
\usage{
db2vegtable(conn, ...)

\method{db2vegtable}{PostgreSQLConnection}(
  conn,
  database,
  description,
  geometry = "plot_centroid",
  header_cols,
  simplify_header = TRUE,
  as_list = FALSE,
  ...
)
}
\arguments{
\item{conn}{A database connection provided by \code{\link[=dbConnect]{dbConnect()}}.}

\item{...}{Further arguments passed to \code{\link[=db2taxlist]{db2taxlist()}}.}

\item{database}{A character value indicating the name of the database to be
imported.}

\item{description}{Named vector with metadata.}

\item{geometry}{Name of the variable in header containing the geometry of
the plots.}

\item{header_cols}{A character vector with the names of the columns required.}

\item{simplify_header}{A logical value indicating whether empty columns
should be deleted from slot header or not.}

\item{as_list}{Logical value indicating whether a list or an object of class
\linkS4class{vegtable} should be returned.}
}
\description{
Import and adaption of Postgres tables into objects of class
\linkS4class{vegtable}.

In the case that some schemas are not mentioned, the function assumes such
tables are located in the same schema as the table header. Thus for
databases placed in just one schema, this need to be set only in argument
\code{header_schema}.
}
