% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbSendQuery.R
\name{dbSendQuery}
\alias{dbSendQuery}
\alias{dbSendQuery,PostgreSQLConnection,sql-method}
\title{Send multiple queries to a database}
\usage{
\S4method{dbSendQuery}{PostgreSQLConnection,sql}(conn, statement, ...)
}
\arguments{
\item{conn}{A connection as \linkS4class{PostgreSQLConnection} object.}

\item{statement}{A sql object.}

\item{...}{Further arguments passed among methods.}
}
\description{
A wrapper of \code{\link[DBI:dbSendQuery]{DBI::dbSendQuery()}} for convenience to send SQL commands read
by \code{\link[=read_sql]{read_sql()}}.
Each statement will be executed in a loop.
}
\author{
Miguel Alvarez \email{kamapu78@gmail.com}
}
