% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do_backup.R
\name{do_backup}
\alias{do_backup}
\alias{do_restore}
\title{Create a backup by 'pg_dump' and restore database}
\usage{
do_backup(
  dbname = "",
  host = "localhost",
  port = "5432",
  user = "",
  password = "",
  filepath = getwd(),
  filename = dbname,
  fext = ".backup",
  path_psql = "/usr/bin",
  f_timestamp = "\%Y\%m\%d",
  ...
)

do_restore(
  dbname = "",
  backup,
  host = "localhost",
  port = "5432",
  user = "",
  password = "",
  filepath = getwd(),
  filename = dbname,
  fext = ".backup",
  path_psql = "/usr/bin",
  f_timestamp = "\%Y\%m\%d",
  opts = "--clean",
  ...
)
}
\arguments{
\item{dbname}{Character value indicating the name of the target PostgreSQL
database.}

\item{host}{Character value, the host name.}

\item{port}{Integer value, the port applied for database connection.}

\item{user}{Character value indicating the database user in PostgreSQL.}

\item{password}{Character value, the user's password in PostgreSQL.}

\item{filepath}{Character value indicating the path to the folder, where the
backup file will be restored. It have to be a path relative to the
working directory.}

\item{filename}{Character value, the name of the backup file. If not provided,
the function guess the database's name and a time stamp, where the newest
backup, if many in the same folder, will be used.}

\item{fext}{Character value indicating the extension used for the backup
file, including the leading dot. The extension ".backup" is used by
default.}

\item{path_psql}{Character value indicating the system path to PostgreSQL
binaries.}

\item{f_timestamp}{Character value indicating the format of the timestamp
applied to the backup's name (see \code{\link[=strptime]{strptime()}}).}

\item{...}{Further arguments passed to \code{\link[=system]{system()}}.}

\item{backup}{A character value indicating a specific backup file to be used
for restore.}

\item{opts}{A character value indicating additional options for "pg_restore",
for instance "--clean" or "--create".}
}
\description{
Create a backup by 'pg_dump' and restore database
}
\details{
The function \code{do_backup()} is a wrapper for \strong{pg_dump} and is creating
writing a dump file named by the backed up database, a time stamp (date of
creation) and a suffix (in the case of a further backup done at the same
day).
Naming the files is done in the same way as by \code{\link[=backup_object]{backup_object()}}.

The function \code{do_restore()} can restore the created backup files through
\strong{pg_restore}.
If several backups have been accumulated in a common folder, this function is
able to recognize the newest one in the same way as \code{\link[=load_last]{load_last()}}.
}
\author{
Miguel Alvarez
}
