% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_badge.R
\name{draw_badge}
\alias{draw_badge}
\title{Draw a GIVD or a version badge}
\usage{
draw_badge(
  id,
  path,
  version = Sys.Date(),
  link,
  type = "version",
  encoding = "UTF-8",
  ...
)
}
\arguments{
\item{id}{A character value indicating the ID of the database in GIVD.}

\item{path}{A character value indicatin the path where to write the svg file,
including the file name.}

\item{version}{A character value for the version, usually the date of the
version.}

\item{link}{A character value with a url to link the version badge (usually
a page for the database).}

\item{type}{The type of the badge, either "version" or "givd".}

\item{encoding}{Character value indicating the encoding of the output file,
passed to \code{\link[=file]{file()}}.}

\item{...}{Further arguments passed to \code{\link[=file]{file()}}.}
}
\description{
Badges to insert in database pages for displaying the ID and link to the
respective entry at \href{https://www.givd.info/}{GIVD} (Global Index of
Vegetation-Plot Databases).

This function writes the respective svg badge and suggests the markdown code
needed to insert the badge in your site. Note that the markdown code may need
edition.
}
