% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_list.R
\name{file_list}
\alias{file_list}
\alias{file_list.lib_df}
\alias{file_list,lib_df-method}
\alias{file_list2string}
\alias{file_list2string.data.frame}
\alias{file_list2string,data.frame-method}
\alias{file_list2string.lib_db}
\alias{file_list2string,lib_db-method}
\title{Description of files associated to entries in BibTeX databases}
\usage{
file_list(x, ...)

\method{file_list}{lib_df}(x, ...)

file_list2string(x, ...)

\method{file_list2string}{data.frame}(x, ...)

\method{file_list2string}{lib_db}(x, ...)
}
\arguments{
\item{x}{Either objects of class \linkS4class{lib_df}, \linkS4class{lib_db},
or a data.frame, depending on function and method.
The function \code{file_list()} applied to a \linkS4class{lib_df} object retrieves
a data frame with the details of linked documents. The function
\code{file_list2string()} will do the back conversion from a data frame to
the string required by \strong{JabRef}.}

\item{...}{Further parameters passed among methods.}
}
\value{
For \code{file_list()} applied to a \linkS4class{lib_df} object, a data frame.
For \code{file_list2string()} applied to a data frame, a data frame with the
columns \strong{bibtexkey} (the reference identifier) and \strong{file} (the respective
string with linked files).
}
\description{
Main documents as well as supplementary material may be linked to the
respective bibliographic entries in a BibTeX database. File names, MIME-types
and descriptions are stored in a single column by JabRef, separating columns
by colons and multiple files by semicolons. In a relational database, file
information may be stored in a separated table and linked by bibtexkeys.
}
