% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fix_double.R
\name{fix_double}
\alias{fix_double}
\alias{fix_double.PostgreSQLConnection}
\alias{fix_double,PostgreSQLConnection-method}
\title{Combine names across data sets}
\usage{
fix_double(conn, ...)

\method{fix_double}{PostgreSQLConnection}(conn, new_id, old_id, ...)
}
\arguments{
\item{conn}{Connection to the database as \linkS4class{PostgreSQLConnection}.}

\item{...}{Further arguments passed among methods.}

\item{new_id}{Vector with IDs of the correct name.}

\item{old_id}{Vector with IDs of the names that have to be replaced and
deleted.}
}
\description{
When names inserted with no authors or with misspelled authors, duplicated
entries will be created, which may cause conflicts if one name is corrected
since the database does not allow duplicated combinations of names and
authors.

In this routines one name (parameter \code{'old_id'}) will be replaced by another
alternative name (parameter \code{'new_id'}) in all usages and the first will be
finally deleted from the database.

If the two alternatives are already used within the same taxonomy (another
constraint of the database), this function retrieves an error message.
}
