% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_countries.R
\name{get_countries}
\alias{get_countries}
\alias{get_countries,PostgreSQLConnection,vegtable-method}
\title{After import assignment of countries for vegtable objects}
\usage{
get_countries(conn, object, ...)

\S4method{get_countries}{PostgreSQLConnection,vegtable}(
  conn,
  object,
  rel_name = "country_code",
  geom_name = "plot_centroid",
  ...
)
}
\arguments{
\item{conn}{A \linkS4class{PostgreSQLConnection} object.}

\item{object}{A \linkS4class{vegtable} object.}

\item{...}{Further arguments passed among methods.}

\item{rel_name}{A character value indicating the name of the new inserted
relation.}

\item{geom_name}{A character value indicating the name of the geometry
including plot coordinates. It is assumed to be a relation in the object.}
}
\value{
An object of class \linkS4class{vegtable} with added country codes.
}
\description{
This function access to a map in the database and assign countries of
samples by overlying locations with the map.

NA's are re-assigned by proximity.
}
\author{
Miguel Alvarez \email{malvarez@uni-bonn.de}
}
