% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_description.R
\name{get_description}
\alias{get_description}
\alias{get_description.PostgreSQLConnection}
\alias{get_description,PostgreSQLConnection-method}
\title{Get descriptions of table columns from Postgres tables}
\usage{
get_description(conn, ...)

\method{get_description}{PostgreSQLConnection}(conn, ...)
}
\arguments{
\item{conn}{A \linkS4class{PostgreSQLConnection} object.}

\item{...}{Further arguments passed among methods.}
}
\value{
An object of class \link{data.frame}.
}
\description{
Descriptions of variables stored in PostgreSQL tables.

This function produces a data frame containing information on schemas,
tables and descriptions (i.e. comments) for every single column in tables
stored at a PostgreSQL database.
}
\author{
Miguel Alvarez \email{malvarez@uni-bonn.de}
}
