% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_precision.R
\name{get_precision}
\alias{get_precision}
\title{Get length and precision for numeric variables}
\usage{
get_precision(x)
}
\arguments{
\item{x}{A numeric vector.}
}
\description{
This function can be used for define data types in databases according to
vector properties (length and precision).

Get precision to define numeric characters in databases (e.g.
\code{NUMERIC(5,3)}).
}
\examples{
data(iris)
get_precision(iris$Sepal.Length)
}
\author{
Miguel Alvarez \email{kamapu78@gmail.com}
}
