% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insert_concept.R
\name{insert_concept}
\alias{insert_concept}
\alias{insert_concept.PostgreSQLConnection}
\alias{insert_concept,PostgreSQLConnection-method}
\title{Insert names or concepts in PostgreSQL taxonomic lists}
\usage{
insert_concept(conn, ...)

\method{insert_concept}{PostgreSQLConnection}(conn, taxonomy, df, clean = TRUE, ...)
}
\arguments{
\item{conn}{A database connection provided by \code{\link[=dbConnect]{dbConnect()}}.}

\item{...}{Further arguments passed among methods.}

\item{taxonomy}{Character value with the name of the taxonomy in the
database.}

\item{df}{A data frame with new names and related information. Two columns
are mandatory, namely \strong{usage_name} and \strong{author_name}, both as
character vectors.}

\item{clean}{A logical value indicating cleaning of characters.}
}
\description{
Insert synonyms to existing taxa in a PostgreSQL version of \linkS4class{taxlist}
objects.

This function is updating the tables \code{taxonNames} and \code{names2concepts} in
the PostgreSQL version of the database.
}
