% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insert_synonym.R
\name{insert_synonym}
\alias{insert_synonym}
\alias{insert_synonym.PostgreSQLConnection}
\alias{insert_synonym,PostgreSQLConnection-method}
\title{Insert synonym to existing taxon concepts in database}
\usage{
insert_synonym(conn, ...)

\method{insert_synonym}{PostgreSQLConnection}(
  conn,
  taxonomy,
  schema = "plant_taxonomy",
  df,
  clean = TRUE,
  ...
)
}
\arguments{
\item{conn}{A database connection provided by \code{\link[=dbConnect]{dbConnect()}}.}

\item{...}{Further arguments passed among methods.}

\item{taxonomy}{Character value with the name of the taxonomy in the
database.}

\item{schema}{Character value indicating the name of the schema containing
all taxonomic tables in the database.}

\item{df}{A data frame with new names and related information. Two columns
are mandatory, namely \strong{taxon_concept_id}, \strong{usage_name} and
\strong{author_name}, both as character vectors.}

\item{clean}{A logical value indicating cleaning of characters.}
}
\description{
Adding synonyms to existing concepts in database.
}
