% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib_db-class.R
\docType{class}
\name{lib_db-class}
\alias{lib_db-class}
\alias{lib_db}
\title{Class lib_db.}
\description{
A class created as intermediate format between \linkS4class{lib_df} objects and a
database.
}
\section{Slots}{

\describe{
\item{\code{dir}}{A list including the folder containing files and a connection to
a database.}

\item{\code{main_table}}{A \linkS4class{lib_df} object.}

\item{\code{file_list}}{A data frame with a list of files associated to a library
entry.}

\item{\code{relations}}{A list with data frames related to single columns in
main_table.}
}}

\examples{
showClass("lib_db")

}
