% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_names.R
\name{merge_names}
\alias{merge_names}
\alias{merge_names.PostgreSQLConnection}
\alias{merge_names,PostgreSQLConnection-method}
\title{Replace names across taxonomies}
\usage{
merge_names(conn, ...)

\method{merge_names}{PostgreSQLConnection}(
  conn,
  usage_id,
  schema = "plant_taxonomy",
  relations = NULL,
  eval = TRUE,
  ...
)
}
\arguments{
\item{conn}{Connection to the database as \linkS4class{PostgreSQLConnection}.}

\item{...}{Further arguments passed among methods.}

\item{usage_id}{A vector of Identifiers for the name. The first value will be
used as the new name and any other will be replaced by the first.
Note that at the end of this process all names but the first will be
deleted.}

\item{schema}{A character value indicating the name of the schema containing
taxonomies in the database.}

\item{relations}{Either a data frame with the columns \strong{schema}, \strong{table}
and \strong{column} or a list of character vectors including the name of
schema, table, and column, where the names need to be changed.}

\item{eval}{A logical value, whether the function will execute the resulting
sql or not.}
}
\value{
An invisible \link[divDB:sql-class]{divDB::sql} object.
}
\description{
Correcting or completing names within a taxonomic list may cause
duplicated names in a common list of names. This situation is not allowed
by the used Postgres model. Thus, you may requery to replace occurrences of a
name with existing entries.

The raplacement need to consider potential issues caused by inserting the new
name and do updates in multiple tables.
}
