% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_taxa.R
\name{merge_taxa}
\alias{merge_taxa}
\alias{merge_taxa,PostgreSQLConnection,numeric,missing-method}
\title{Merge multiple concepts into one}
\usage{
\S4method{merge_taxa}{PostgreSQLConnection,numeric,missing}(
  object,
  concepts,
  names2concepts,
  taxon_relations,
  taxon_traits,
  ...
)
}
\arguments{
\item{object}{A database connection provided by \code{\link[=dbConnect]{dbConnect()}}.}

\item{concepts}{ID of taxon concepts to be merged.}

\item{taxon_relations, taxon_traits, names2concepts}{Character vectors
containing the name of the schema and for the respective information.}

\item{...}{Further arguments passed among methods.}
}
\description{
Different taxon concepts may be merged by taxonomic revisions.

All concepts indicated in argument \code{concept_id} will be set as
synonyms of the first concept in the vector.

Take care of producing backups before starting the manipulation of
databases.
}
\author{
Miguel Alvarez \email{kamapu78@gmail.com}
}
