% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_taxonomy.R
\name{new_taxonomy}
\alias{new_taxonomy}
\alias{new_taxonomy,PostgreSQLConnection,taxlist-method}
\title{Create new database from a taxlist object}
\usage{
new_taxonomy(conn, obj, ...)

\S4method{new_taxonomy}{PostgreSQLConnection,taxlist}(conn, obj, taxonomy, schema, schema_refs, ...)
}
\arguments{
\item{conn}{A \linkS4class{PostgreSQLConnection} connecting to a database, where
the new list will be stored.}

\item{obj}{A \linkS4class{taxlist} object containing the new database.}

\item{...}{Further arguments passed among methods (not in use).}

\item{taxonomy}{A character value indicating the name (ID) of the new
taxonomy in the database.}

\item{schema}{A character value with the name of the new schema containing
the taxonomic list.}

\item{schema_refs}{A character value with the name of the new schema
containing the references used as taxon views. It is recommended to have
a separated schema for the references, especially if the views are
formated as a BibTeX database (\linkS4class{lib_df} object).}
}
\description{
Create a new schema in a connected database containing a taxonomic list from
a taxlist object.
}
