% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pg_insert_observations.R
\name{pg_insert_observations}
\alias{pg_insert_observations}
\alias{pg_insert_observations.PostgreSQLConnection}
\alias{pg_insert_observations,PostgreSQLConnection-method}
\title{Insert plots and samples in PostgreSQL}
\usage{
pg_insert_observations(conn, ...)

\method{pg_insert_observations}{PostgreSQLConnection}(
  conn,
  samples,
  header,
  names2concepts,
  db_samples,
  db_header,
  geom,
  ...
)
}
\arguments{
\item{conn}{A database connection provided by \code{\link[DBI]{dbConnect}}.}

\item{...}{Further arguments (not yet in use).}

\item{samples}{A data frame including records of species in plots. Columns
'ReleveID' and 'TaxonUsageID' are mandatory in this object.}

\item{header}{An object of class \link[sp:SpatialPointsDataFrame-class]{sp::SpatialPointsDataFrame} including
the information on the plots. In the attribute table the columns 'ReleveID'
and 'original_number' are mandatory.}

\item{names2concepts, db_samples, db_header}{Character vectors including the
name of the respective schema and table in database.}

\item{geom}{Character value indicating the name of the geometry variable in
the database.}
}
\description{
Plots and records of species in the plots will be jointly inserted in a
PostgreSQL database, whereby a series of checks will be carried out previous
to insert the data.

This function carry out a series of checks on the input data previous to
import them in the database (if not errored). The variable 'original_number'
will be used for cross-check identities of entries in samples and in header.
Connection to taxa via 'TaxonUsageID'.
}
\author{
Miguel Alvarez \email{kamapu78@gmail.com}
}
