% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print-methods.R
\docType{methods}
\name{print-methods}
\alias{print-methods}
\alias{print.comp_db}
\title{Print comparisons of multiple data frames}
\usage{
\method{print}{comp_db}(x, ...)
}
\arguments{
\item{x}{An object of class \linkS4class{comp_df2}.}

\item{...}{Further arguments passed among methods.}
}
\description{
Comparing multiple pairs of data frames or \linkS4class{lib_df} objects can be
assigned to a single list setting as class
\code{class(x) <- c("comp_df2", "list")}.
}
