% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_concepts.R
\name{query_concepts}
\alias{query_concepts}
\alias{query_concepts.PostgreSQLConnection}
\alias{query_concepts,PostgreSQLConnection-method}
\title{Query taxon concepts}
\usage{
query_concepts(conn, ...)

\method{query_concepts}{PostgreSQLConnection}(
  conn,
  query,
  taxonomy,
  schema = "plant_taxonomy",
  case = FALSE,
  exact = FALSE,
  ...
)
}
\arguments{
\item{conn}{Connection to the database as \linkS4class{PostgreSQLConnection}.}

\item{...}{Further arguments passed to \code{\link[=db2taxlist]{db2taxlist()}}.}

\item{query}{A character value that will be matched with the names stored in
the database.}

\item{taxonomy}{Character value with the name of the taxonomy in the
database.}

\item{schema}{Character value indicating the name of the schema containing
taxonomic information within the database.}

\item{case}{A logical value indicating whether the match should be case
sensitive (\code{TRUE}) or not (\code{FALSE}, the default).}

\item{exact}{A logical value indicating whether the exact string is queried.}
}
\value{
A \linkS4class{taxlist} object with the queried taxon concepts.
}
\description{
Extract taxonomies by matching character strings.
}
