% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_name.R
\name{query_name}
\alias{query_name}
\alias{query_name.PostgreSQLConnection}
\alias{query_name,PostgreSQLConnection-method}
\title{Query taxon usage names}
\usage{
query_name(conn, ...)

\method{query_name}{PostgreSQLConnection}(conn, query, case = FALSE, concepts = FALSE, accepted = FALSE, ...)
}
\arguments{
\item{conn}{Connection to the database as \linkS4class{PostgreSQLConnection}.}

\item{...}{Further arguments passed among methods.}

\item{query}{A character value that will be matched with the names stored in
the database.}

\item{case}{A logical value indicating whether the match should be case
sensitive (\code{TRUE}) or not (\code{FALSE}, the default).}

\item{concepts}{A logical value indicating whether taxon concepts should be
displayed (\code{TRUE}) or just the names (\code{FALSE}, the default).}

\item{accepted}{A logical value indicating whether the respective accepted
names should be displayed (\code{TRUE}) or not (\code{FALSE}, the default). This
works only if \code{'concepts = TRUE'}.}
}
\description{
A quick display of stored taxon names by matching them with a character
string.

This function can also provide details about the associated taxon concpets,
taxonomies and accepted names for each taxonomy.
}
