% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_names.R
\name{query_names}
\alias{query_names}
\alias{query_names.PostgreSQLConnection}
\alias{query_names,PostgreSQLConnection-method}
\title{Query taxon usage names}
\usage{
query_names(conn, ...)

\method{query_names}{PostgreSQLConnection}(
  conn,
  query,
  schema = "plant_taxonomy",
  case = FALSE,
  exact = FALSE,
  concepts = FALSE,
  accepted = FALSE,
  ...
)
}
\arguments{
\item{conn}{Connection to the database as
\link[RPostgreSQL:PostgreSQLConnection-class]{RPostgreSQL::PostgreSQLConnection}.}

\item{...}{Further arguments passed among methods.}

\item{query}{A character value that will be matched with the names stored in
the database.}

\item{schema}{Character value indicating the name of the schema containing
taxonomic information within the database.}

\item{case}{A logical value indicating whether the match should be case
sensitive (\code{TRUE}) or not (\code{FALSE}, the default). I does not apply if
\code{'exact = TRUE'}.}

\item{exact}{A logical value indicating whether the exact string is queried.}

\item{concepts}{A logical value indicating whether taxon concepts should be
displayed (\code{TRUE}) or just the names (\code{FALSE}, the default).}

\item{accepted}{A logical value indicating whether the respective accepted
names should be displayed (\code{TRUE}) or not (\code{FALSE}, the default). This
works only if \code{'concepts = TRUE'}.}
}
\description{
A quick display of stored taxon names by matching them with a character
string.

This function can also provide details about the associated taxon concpets,
taxonomies and accepted names for each taxonomy.
}
