% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_sql.R
\name{read_sql}
\alias{read_sql}
\title{Read SQL scripts and split it in SQL statements}
\usage{
read_sql(file, end = ";", comment = "--", ...)
}
\arguments{
\item{file}{A character value indicating the path to the script that have to
be read.}

\item{end}{Symbol set at the end of a statement (semicolon by default).}

\item{comment}{Symbol used to start a comment in script (two dashes by
default).}

\item{...}{Further arguments passed to \code{\link[=readLines]{readLines()}}.}
}
\description{
Read SQL scripts and split it in SQL statements
}
\details{
SQL scripts may mix SQL statements with comments. This function extract the
respective statements and split them into a list, where each element is a
statement.
}
\author{
Miguel Alvarez
}
