% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql-class.R
\name{sql-class}
\alias{sql-class}
\title{A class for lists of SQL commands}
\description{
Multiple SQL commands can be read from a *.sql file by \code{\link[=read_sql]{read_sql()}} and
stored as a character vector.
}
