% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db2taxlist.R
\name{db2taxlist}
\alias{db2taxlist}
\alias{db2taxlist.PostgreSQLConnection}
\alias{db2taxlist,PostgreSQLConnection-method}
\title{Import relational databases into taxlist objects}
\usage{
db2taxlist(conn, ...)

\method{db2taxlist}{PostgreSQLConnection}(
  conn,
  taxonomy,
  concepts,
  schema = "plant_taxonomy",
  schema_refs = "bib_references",
  subset_levels = TRUE,
  keep_parents = FALSE,
  keep_children = FALSE,
  as_list = FALSE,
  ...
)
}
\arguments{
\item{conn}{A database connection provided by \code{\link[divDB:connect_db]{divDB::connect_db()}}.}

\item{...}{Further arguments passed among methods (not used here).}

\item{taxonomy}{Character value with the name of the taxonomy in the
database.}

\item{concepts}{A vector with taxon concept IDs to be included in the output.
If not provided the whole taxonomy will be imported. IDs belonging to a
different taxonomy will cause an error message.}

\item{schema}{Character value indicating the name of the schema containing
taxonomic information within the database.}

\item{schema_refs}{Character value indicating the name of the schema
containing the electronic library with taxon views.}

\item{subset_levels}{Logical value indicating whether taxonomic ranks should
be restricted to the used ones or all ranks available in the database.}

\item{keep_parents}{A logical value indicating whether parents of queried
concepts should be included in the output or not. It works only if an
argument is provided for the parameter \code{'concepts'}.}

\item{keep_children}{A logical value indicating whether children of queried
concepts should be included in the output or not. It works only if an
argument is provided for the parameter \code{'concepts'}.}

\item{as_list}{Logical value indicating whether the output should be a list
or a \link[taxlist:taxlist-class]{taxlist::taxlist} object.}
}
\description{
Ad-hoc function for importing Postgres tables into objects of class
\link[taxlist:taxlist-class]{taxlist::taxlist}.

This function has been modified to a new version of the database.
}
