% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_description.R
\name{get_description}
\alias{get_description}
\alias{get_description.PostgreSQLConnection}
\alias{get_description,PostgreSQLConnection-method}
\title{Produce a table of all variables from a database}
\usage{
get_description(conn, ...)

\method{get_description}{PostgreSQLConnection}(conn, ...)
}
\arguments{
\item{conn}{A database connection provided by \code{\link[DBI:dbConnect]{DBI::dbConnect()}}.}

\item{...}{Further arguments passed to \code{\link[=dbGetQuery]{dbGetQuery()}}.}
}
\value{
A data frame.
}
\description{
Overview of all variables stored in a PostgreSQL database including the
respective schema and table as well as the descriptions stored as comments.
}
\author{
Miguel Alvarez \email{kamapu78@gmail.com}
}
