% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insert_concept.R
\name{insert_concept}
\alias{insert_concept}
\alias{insert_concept.PostgreSQLConnection}
\alias{insert_concept,PostgreSQLConnection-method}
\title{Insert taxonomic concepts in database}
\usage{
insert_concept(conn, ...)

\method{insert_concept}{PostgreSQLConnection}(
  conn,
  taxonomy,
  schema = "plant_taxonomy",
  df,
  clean = TRUE,
  ...
)
}
\arguments{
\item{conn}{A database connection provided by \code{\link[=dbConnect]{dbConnect()}}.}

\item{...}{Further arguments passed among methods.}

\item{taxonomy}{Character value with the name of the taxonomy in the
database.}

\item{schema}{Character value indicating the name of the schema containing
all taxonomic tables in the database.}

\item{df}{A data frame with new names and related information. Two columns
are mandatory, namely \strong{usage_name} and \strong{author_name}, both as
character vectors.}

\item{clean}{A logical value indicating cleaning of characters.}
}
\description{
Insert new taxonomic concepts by using accepted names.
A previous use of the name in the taxonomy will retrieve an error message.
Names that are new for the database will be inserted in the main list of
names.
}
