% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insert_concept.R
\name{insert_concept}
\alias{insert_concept}
\alias{insert_concept.PostgreSQLConnection}
\alias{insert_concept,PostgreSQLConnection-method}
\alias{insert_concept_swea}
\alias{insert_concept_swea.PostgreSQLConnection}
\alias{insert_concept_swea,PostgreSQLConnection-method}
\title{Insert names or concepts in PostgreSQL taxonomic lists}
\usage{
insert_concept(conn, ...)

\method{insert_concept}{PostgreSQLConnection}(
  conn,
  taxon_names,
  taxon_relations,
  names2concepts,
  taxon_views,
  taxon_levels,
  df,
  clean = TRUE,
  ...
)

insert_concept_swea(conn, ...)

\method{insert_concept_swea}{PostgreSQLConnection}(
  conn,
  taxon_names = c("tax_commons", "taxonNames"),
  taxon_relations = c("swea_dataveg", "taxonRelations"),
  names2concepts = c("swea_dataveg", "names2concepts"),
  taxon_views = c("bib_references", "main_table"),
  taxon_levels = c("tax_commons", "taxonLevels"),
  df,
  ...
)
}
\arguments{
\item{conn}{A database connection provided by \code{\link[=dbConnect]{dbConnect()}}.}

\item{...}{Further arguments passed among methods.}

\item{taxon_names, taxon_relations, names2concepts, taxon_views, taxon_levels}{Character vectors indicating the name of the respective schemas and
tables in database.}

\item{df}{A data frame with new names and related information (including
taxon concept ID).}

\item{clean}{A logical value indicating cleaning of characters.}
}
\description{
Insert synonyms to existing taxa in a PostgreSQL version of \linkS4class{taxlist}
objects.

This function is updating the tables \code{taxonNames} and \code{names2concepts} in
the PostgreSQL version of the database.
}
