% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insert_names.R
\name{insert_names}
\alias{insert_names}
\alias{insert_names,PostgreSQLConnection,data.frame,character-method}
\alias{insert_names,PostgreSQL,data.frame,character-method}
\alias{insert_names,PostgreSQLConnection,data.frame,missing-method}
\alias{insert_names,PostgreSQL,data.frame,missing-method}
\title{Insert new taxonomic names in database}
\usage{
insert_names(conn, df, schema, ...)

\S4method{insert_names}{PostgreSQLConnection,data.frame,character}(conn, df, schema, clean = TRUE, eval = TRUE, update = FALSE, ...)

\S4method{insert_names}{PostgreSQLConnection,data.frame,missing}(conn, df, schema, ...)
}
\arguments{
\item{conn}{A \link[RPostgreSQL:PostgreSQLConnection-class]{RPostgreSQL::PostgreSQLConnection} object connecting to
a target database.}

\item{df}{A data frame with the list of names that will be inserted to the
database. Two columns are mandatory in this data frame, namely
\strong{usage_name} and \strong{author_name}.}

\item{schema}{A character value indicating the name of the schema containing
the taxonomic list. If the table \strong{taxon_names} does not exists in this
schema, this function retrieves an error message.}

\item{...}{Further arguments passed among methods (not in use).}

\item{clean}{A logical value indicating cleaning of characters.}

\item{eval}{A logical value indicating whether the produced SQL commands
should be sent to the database or not.}

\item{update}{A logical value indicating whether attributes of existing names
(mentioned as recycled) should be updated according to the input data
frame or not.}
}
\description{
Different routines may require to compare a list of new names with existing
database.
Names that are already in the database will be maintained and only new names
will get inserted in the database.

For existing names, if further information is provided, this will be updated
in the database.
}
