% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insert_synonym.R
\name{insert_synonym}
\alias{insert_synonym}
\alias{insert_synonym.PostgreSQLConnection}
\alias{insert_synonym,PostgreSQLConnection-method}
\title{Insert synonyms in PostgreSQL taxonomic lists}
\usage{
insert_synonym(conn, ...)

\method{insert_synonym}{PostgreSQLConnection}(conn, taxonomy, df, clean = TRUE, ...)
}
\arguments{
\item{conn}{A database connection provided by \code{\link[=dbConnect]{dbConnect()}}.}

\item{...}{Further arguments passed among methods.}

\item{taxonomy}{Character value with the name of the taxonomy in the
database.}

\item{df}{A data frame with new names and related information. Three columns
are mandatory, \strong{usage_name} and \strong{author_name} as character vectors,
and \strong{taxon_concept_id} as integer.}

\item{clean}{A logical value, whether strings in input 'df' should be cleaned
or not (see \code{\link[=clean_strings]{clean_strings()}}).}
}
\description{
Insert synonyms to existing taxa in a PostgreSQL version of \linkS4class{taxlist}
objects.

This function is updating the tables \code{taxonNames} and \code{names2concepts} in
the PostgreSQL version of the database.
}
\author{
Miguel Alvarez \email{kamapu78@gmail.com}
}
