% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insert_synonym.R
\name{insert_synonym}
\alias{insert_synonym}
\alias{insert_synonym.PostgreSQLConnection}
\alias{insert_synonym,PostgreSQLConnection-method}
\alias{insert_synonym_swea}
\title{Insert synonyms in PostgreSQL taxonomic lists}
\usage{
insert_synonym(conn, ...)

\method{insert_synonym}{PostgreSQLConnection}(
  conn,
  taxon_names,
  taxon_relations,
  names2concepts,
  df,
  clean = TRUE,
  ...
)

insert_synonym_swea(
  conn,
  taxon_names = c("tax_commons", "taxonNames"),
  taxon_relations = c("swea_dataveg", "taxonRelations"),
  names2concepts = c("swea_dataveg", "names2concepts"),
  df,
  ...
)
}
\arguments{
\item{conn}{A database connection provided by \code{\link[=dbConnect]{dbConnect()}}.}

\item{...}{Further arguments passed among methods.}

\item{taxon_names}{A character vector of length 2 indicating the name of the
respecitve schema and table in Postgres.}

\item{taxon_relations}{taxon_names A character vector of length 2 indicating
the name of the respecitve schema and table in Postgres.}

\item{names2concepts}{taxon_names A character vector of length 2 indicating
the name of the respecitve schema and table in Postgres.}

\item{df}{A data frame with new names and related information (including
taxon concept ID).}

\item{clean}{A logical value, whether strings in input 'df' should be cleaned
or not (see \code{\link[=clean_strings]{clean_strings()}}).}
}
\description{
Insert synonyms to existing taxa in a PostgreSQL version of \linkS4class{taxlist}
objects.

This function is updating the tables \code{taxonNames} and \code{names2concepts} in
the PostgreSQL version of the database.
}
\author{
Miguel Alvarez \email{kamapu78@gmail.com}
}
