% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insert_synonyms.R
\name{insert_synonyms}
\alias{insert_synonyms}
\alias{insert_synonyms.PostgreSQLConnection}
\alias{insert_synonyms,PostgreSQLConnection-method}
\title{Insert synonym to existing taxon concepts in database}
\usage{
insert_synonyms(conn, ...)

\method{insert_synonyms}{PostgreSQLConnection}(conn, df, schema = "plant_taxonomy", clean = TRUE, ...)
}
\arguments{
\item{conn}{A database connection provided by \code{\link[=dbConnect]{dbConnect()}}.}

\item{...}{Further arguments passed among methods.}

\item{df}{A data frame with new names and related information. Two columns
are mandatory, namely \strong{taxon_concept_id}, \strong{usage_name} and
\strong{author_name}, both as character vectors.}

\item{schema}{Character value indicating the name of the schema containing
all taxonomic tables in the database.}

\item{clean}{A logical value indicating cleaning of characters.}
}
\description{
Adding synonyms to existing concepts in database.
}
