% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_names.R
\name{merge_names}
\alias{merge_names}
\alias{merge_names,PostgreSQLConnection-method}
\title{Merge names across data sets}
\usage{
merge_names(object, ...)

\S4method{merge_names}{PostgreSQLConnection}(
  object,
  new_id,
  old_id,
  taxon_names = c("tax_commons", "taxonNames"),
  schemas = c("swea_dataveg", "sudamerica"),
  ...
)
}
\arguments{
\item{object}{A database connection established by \code{\link[=dbConnect]{dbConnect()}}.}

\item{...}{Further arguments passed among methods.}

\item{new_id}{Integer value, the persisting name id ('TaxonUsageID').}

\item{old_id}{Integer value, the name id ('TaxonUsageID') to be replaced.}

\item{taxon_names}{Character vector indicating the name of schema and table
containing taxon names.}

\item{schemas}{Character vector indicating the schemas containing the
respective databases sharing taxon names.}
}
\description{
When completing authors names in common names tables, duplicated entries may
emerge.
To merge incomplete entry with a complete one, the replacement have to be
done across all data sets (vegetation-plot databases).
}
